added suggestion to save coords and also reactive text (needs to be prettified)
I added the text output under the map that changes when you click on a
point. It needs to be made prettier, either by looking at the box() for
uioutput in shiny, or just writing own HTML.

I also added some suggestions to maybe once the points have been
geocoded once, we could save the lat/long columns and write back into
the google sheets, so that it doesn’t need to be geocoded each time
someone loads the app?